editar post

+++ src/componentes/seccionPerfil.js
@@ -1,132 +1,5 @@
-import { obtenerUserPosts, eliminarPost } from '../firebase/funcionesFirestore.js';
+import { obtenerUserPosts, eliminarPost, actualizarPost } from '../firebase/funcionesFirestore.js';
 import { validateSessionStorage } from './validaciones.js';
-/* export const contenidoPerfil = () => {
-  const perfilSeccion = document.createElement('section');
-  perfilSeccion.classList.add('cuerpoPerfil');
-  const userData = JSON.parse(sessionStorage.userSession);
-  perfilSeccion.innerHTML = `        
-        <nav class= "barraNavegacionInferior">         
-            <ul>
-                <li class="list">
-                    <a>
-                        <span class="icon">
-                            <img src="imagenes/users-three.png">
-                        </span>                        
-                    </a>
-                </li>
-                <li class="list">
-                    <a href="#/artmuro">
-                        <span class="icon">
-                            <img src="imagenes/house-fill.png">
-                        </span>                        
-                    </a>
-                </li>
-                <li class="list">
-                    <a>
-                        <span class="icon">
-                            <img src="imagenes/ImgUsuario.png">
-                        </span>                        
-                    </a>
-                </li>
-            </ul>              
-        </nav> 
-    
-        <div class="fondo1">
-            <div class="fondoImagenPerfil">
-                <img src="imagenes/ImgDelUsuario.png">
-            </div>  
-            <div class="fondo2">
-                <div class="imgPerfilUsuario">
-                    <img src="imagenes/ImgUsuario.png">
-                </div>
-                
-                <div class="contenidoTextPerfil">
-                    <h2>${userData.username}</h2>
-                    <p>${userData.descripcion}</p>
-                    <p>Soy de ${userData.ubicacion}</p>
-                </div>
-
-                <div class="botonesDelPerfil">
-                    <button class="btnEditar"><a href='#/arteditarperfil'>Editar Perfil</a></button>
-                    <button class="btnInicio"><a href="#/artmuro">Volver a Inicio</a></button>
-                </div>
-            </div>       
-        </div>  
-        <button class="btnEditar"><a href='#/arteditarperfil'>Editar Perfil</a></button>      
-
-        <div class="seccPostUsuario" id="SeccPublicacionesUsuario">
-            <div class="tableroPost">            
-                <div class="usuarioPost">
-                    <div class="imgUsuarioPost">
-                        <img class="imgPost"src="imagenes/ImgUsuario.png">
-                    </div>
-                    <div class="infoUsuarioPost">
-                        <div class="nombreUsuarioPost"><p>Lucía Lopez</p><img src="imagenes/bxs-user-plus 2.png"></div>
-                        <div class="descripcionUsuarioPost"><p>Amante de los animales</p></div>                                  
-                    </div>
-                    <div class="puntosHorizontales">
-                        <figure class="puntos"></figure>
-                        <figure class="puntos middle2"></figure>
-                        <figure class="puntos"></figure>
-                        <ul class="desplegable2" id="menuListaPost">
-                            <li><a><img src="imagenes/edit.png"><span>Editar</span></a></li>
-                            <li><a><img src="imagenes/delete.png"><span>Eliminar</span></a></li>
-                        </ul>
-                    </div>    
-                </div>                
-
-                <div class="estadoCompartido">
-                    <div class="contenidoCompartido">
-                        <p></p>
-                        <img src="imagenes/gatitoFoto.png">
-                    </div>
-                </div>  
-
-                <div class="botonesReaccion">
-                    <img src="imagenes/heartIcono.png">
-                    <img src="imagenes/comentIcono.png">
-                    <img src="imagenes/compartirIcono.png">
-                </div>
-            </div>
-
-            <div class="tableroPost">            
-                <div class="usuarioPost">
-                    <div class="imgUsuarioPost">
-                        <img class="imgPost"src="imagenes/ImgUsuario.png">
-                    </div>
-                    <div class="infoUsuarioPost">
-                        <div class="nombreUsuarioPost"><p>Lucía Lopez</p><img src="imagenes/bxs-user-plus 2.png"></div>
-                        <div class="descripcionUsuarioPost"><p>Amante de los animales</p></div>                                  
-                    </div>
-                    <div class="puntosHorizontales">
-                        <figure class="puntos"></figure>
-                        <figure class="puntos middle2"></figure>
-                        <figure class="puntos"></figure>
-                        <ul class="desplegable2" >
-                            <li><a><img src="imagenes/edit.png"><span>Editar</span></a></li>
-                            <li><a><img src="imagenes/delete.png"><span>Eliminar</span></a></li>
-                        </ul>
-                    </div>    
-                </div>                
-
-                <div class="estadoCompartido">
-                    <div class="contenidoCompartido">
-                        <p>A veces me preguntan: ¿Por qué inviertes todo ese tiempo y dinero hablando de la amabilidad
-                         para con los animales, cuando existe tanta crueldad hacia el hombre?. A lo que yo respondo: 
-                         Estoy trabajando en las raíces.</p>
-                        <img src="">
-                    </div>
-                </div>  
-
-                <div class="botonesReaccion">
-                    <img src="imagenes/heartIcono.png">
-                    <img src="imagenes/comentIcono.png">
-                    <img src="imagenes/compartirIcono.png">
-                </div>
-            </div>
-        </div>`;
-  return perfilSeccion;
-};*/
 
 const subirContainer = (idPost, creadorPost, apodoUser, postTxt, srcImagenPost) => {
     const divPost = document.createElement('div');
@@ -139,11 +12,14 @@ const subirContainer = (idPost, creadorPost, apodoUser, postTxt, srcImagenPost) 
             <div class="nombreUsuarioPost"><p>${creadorPost}</p><img src="imagenes/bxs-user-plus 2.png"></div>
             <div class="descripcionUsuarioPost"><p>${apodoUser}</p></div>            
         </div>
-        <button class="btnDelete"><img src="imagenes/delete.png"></button>
+        <button class="btnEdit"><img src="imagenes/edit.png"></button>
+        <button class="btnDelete"><img src="imagenes/delete.png"></button>       
     </div>
     <div class="estadoCompartido">
         <div class="contenidoCompartido">
-            <p>${postTxt}</p>
+            <p class="postcontent">${postTxt}</p>
+        </div>
+        <div>
             <img src="${srcImagenPost}">
         </div>
     </div>
@@ -153,7 +29,7 @@ const subirContainer = (idPost, creadorPost, apodoUser, postTxt, srcImagenPost) 
         <img src="imagenes/compartirIcono.png">
     </div>
     `;
-return divPost;
+    return divPost;
 };
 
 const rellenarPerfil = async (containerPost) => {
@@ -162,8 +38,8 @@ const rellenarPerfil = async (containerPost) => {
     datosPost.forEach((post) => {
         containerPost.prepend(subirContainer(post.id, userData.username, userData.descripcion, post.publicacion, ''));
     });
-    btnEliminarPost();;
-
+    btnEliminarPost();
+    btnEditarPost();
 };
   
 export const contenidoPerfil = () => {
@@ -176,29 +52,30 @@ export const contenidoPerfil = () => {
     navInferior.classList.add('barraNavegacionInferior');
     navInferior.innerHTML = `
       <ul>
-      <li class="list">
-          <a>
-              <span class="icon">
-                  <img src="imagenes/users-three.png">
-              </span>
-          </a>
-      </li>
-      <li class="list">
-          <a href="#/artmuro">
-              <span class="icon">
-                  <img src="imagenes/house-fill.png">
-              </span>
-          </a>
-      </li>
-      <li class="list">
-          <a href="#/artperfil">
-              <span class="icon">
-                  <img src="imagenes/ImgUsuario.png">
-              </span>
-          </a>
-      </li>
-      </ul>
+        <li class="list">
+            <a>
+                <span class="icon iconCategorias" id="categorias">
+                    <img src="imagenes/users-three.png">
+                </span>
+            </a>
+        </li>
+        <li class="list">
+            <a href="#/artmuro">
+                <span class="icon">
+                    <img src="imagenes/house-fill.png">
+                </span>
+            </a>
+        </li>
+        <li class="list">
+            <a href="#/artperfil">
+                <span class="icon">
+                    <img src="imagenes/ImgUsuario.png">
+                </span>
+            </a>
+        </li>
+      </ul>  
     `;
+   
     const tableroInformacionUsuario = document.createElement('div');    
     tableroInformacionUsuario.classList.add('fondo1');
     tableroInformacionUsuario.innerHTML = `
@@ -212,8 +89,8 @@ export const contenidoPerfil = () => {
             
             <div class="contenidoTextPerfil">
                 <h2>${userData.username}</h2>
-                <p>(${userData.name})</p>
-                <p>"${userData.descripcion}"</p>
+                <p>${userData.name}</p>
+                <p>${userData.descripcion}</p>
                 <p>${userData.ubicacion}</p>
             </div>
 
@@ -221,13 +98,15 @@ export const contenidoPerfil = () => {
                 <button class="btnEditar"><a href='#/arteditarperfil'>Editar Perfil</a></button>
                 <button class="btnInicio"><a href="#/artmuro">Volver a Inicio</a></button>
             </div>
-        </div>
+        </div>        
     `;
+
     const btnEditarPerfilResponsive = document.createElement('button');    
     btnEditarPerfilResponsive.classList.add('btnEditar');
     btnEditarPerfilResponsive.innerHTML = `
-        <a href='#/arteditarperfil'>Editar Perfil</a>
+        <a href='#/arteditarperfil'>Editar Perfil</a>     
     `;
+
     const contenedorPublicacionesPerfil = document.createElement('div');
     contenedorPublicacionesPerfil.classList.add('seccPostUsuario');
     contenedorPublicacionesPerfil.setAttribute('id', 'SeccPublicacionesUsuario');
@@ -238,22 +117,78 @@ export const contenidoPerfil = () => {
     perfilSeccion.appendChild(tableroInformacionUsuario);
     perfilSeccion.appendChild(btnEditarPerfilResponsive);
     perfilSeccion.appendChild(contenedorPublicacionesPerfil);
+    
     return perfilSeccion;
 };
 
 export const btnEliminarPost = () => {
     const postsCards = document.getElementsByClassName("usuarioPost");
-    //console.log(postsCards);
     Array.from(postsCards).forEach((postCard) => {
         const btnEliminar = postCard.querySelector(".btnDelete");       
         btnEliminar.addEventListener("click", async () => {
+            const confirmarcion = window.confirm("¿Esta seguro que quiere eliminar el post?");
+            if (!confirmarcion) {
+                return
+            }
             const postEliminado = document.getElementById(postCard.id);
             await eliminarPost(postCard.id);  
             console.log("si elimino el post");          
-            postEliminado.parentElement.remove();
-          
-        })
-        
-    })
-   
+            postEliminado.parentElement.remove(); 
+
+        });        
+    });   
 };
+
+export const btnEditarPost = () => {
+    const postsCards = document.getElementsByClassName("tableroPost");
+    Array.from(postsCards).forEach((postCard) => {
+        const btnPencil = postCard.querySelector(".btnEdit");       
+        btnPencil.addEventListener("click", async () => {
+            editarPost(postCard);
+        });
+    });
+}
+
+const editarPost = (postCard) => {
+    
+    const formularioEditar = document.createElement('form');
+    formularioEditar.classList.add('editForm');
+    formularioEditar.innerHTML = `
+        <textarea id="inputEditar" name="inputEditar" rows="5" cols="33"></textarea>    
+        <div class="">
+            <button class="botonGuardarCambios">Guardar</button>
+            <button class="botonCancelarCambios">Cancelar</button>  
+        </div>     
+    `;  
+    const botonGuardarCambios = formularioEditar.querySelector(".botonGuardarCambios");
+    const botonCancelarCambios = formularioEditar.querySelector(".botonCancelarCambios");
+    const estadoCompartido = postCard.querySelector(".estadoCompartido");
+
+    const contenidoCompartido = estadoCompartido.firstElementChild;
+    formularioEditar.firstElementChild.value = contenidoCompartido.firstElementChild.textContent;
+    
+    // borramos el texto y agregamos el formulario de edicion
+    estadoCompartido.innerHTML = "";
+    estadoCompartido.appendChild(formularioEditar);
+
+    botonCancelarCambios.addEventListener('click', (e) => {
+        e.preventDefault();
+        estadoCompartido.innerHTML = "";
+        estadoCompartido.appendChild(contenidoCompartido);
+    });
+
+    botonGuardarCambios.addEventListener('click', (e) => {
+        e.preventDefault();
+        const postId = postCard.firstElementChild.id;
+        const form = e.target.parentElement.parentElement;
+        const changedText = form.firstElementChild.value;
+        actualizarPost(postId, changedText)
+            .then(() => {
+                console.log("sabemos si funciono");
+                contenidoCompartido.firstElementChild.textContent = changedText;
+                estadoCompartido.innerHTML = ""
+                estadoCompartido.appendChild(contenidoCompartido);
+            });
+    });
+}
+
